+++ Sources/MapboxNavigation/NavigationViewportDataSource.swift
@@ -345,13 +345,13 @@ public class NavigationViewportDataSource: ViewportDataSource {
             if options.followingCameraOptions.paddingUpdatesAllowed || followingMobileCamera.padding == nil {
                 followingMobileCamera.padding = UIEdgeInsets(top: followingMobileCameraAnchor.y,
                                                              left: viewportPadding.left,
-                                                             bottom: UIScreen.main.bounds.height - followingMobileCameraAnchor.y + 1.0,
+                                                             bottom: mapView.bounds.height - followingMobileCameraAnchor.y + 1.0,
                                                              right: viewportPadding.right)
                 
-                if let mainCarPlayScreen = UIScreen.mainCarPlay {
+                if mapView.window?.screen.traitCollection.userInterfaceIdiom == .carPlay {
                     followingCarPlayCamera.padding = UIEdgeInsets(top: followingCarPlayCameraAnchor.y,
                                                                   left: carPlayCameraPadding.left,
-                                                                  bottom: mainCarPlayScreen.bounds.height - followingCarPlayCameraAnchor.y + 1.0,
+                                                                  bottom: mapView.bounds.height - followingCarPlayCameraAnchor.y + 1.0,
                                                                   right: carPlayCameraPadding.right)
                 } else {
                     followingCarPlayCamera.padding = carPlayCameraPadding